+++ app/Http/Livewire/Configs/ShowConfigs.php
@@ -3,6 +3,7 @@
 namespace App\Http\Livewire\Configs;
 
 use App\Models\Grid;
+use App\Models\Bot;
 use Livewire\Component;
 use Livewire\WithPagination;
 
@@ -40,10 +41,18 @@ class ShowConfigs extends Component
     public function destroy()
     {
         if ($this->deleteId > 0) {
-            $record = Grid::find($this->deleteId);
-            if(auth()->user()->id == $record->user_id){
-                $record->delete();
-                session()->flash('message', 'Configuration successfully deleted.');
+            $bots_using_grid_count = Bot::where('grid_id', $this->deleteId)->count();
+            if ($bots_using_grid_count == 0) {
+                $record = Grid::find($this->deleteId);
+                if(auth()->user()->id == $record->user_id){
+                    $record->delete();
+                    session()->flash('message', 'Configuration successfully deleted.');
+                }
+            } else {
+                $this->dispatchBrowserEvent('alert',[
+                    'type' => 'error',
+                    'message' => "Can not delete grid that its used by a bot."
+                ]);
             }
             $this->deleteId = 0;
         }
